requirements.txt
click
dataclasses
fastapi
h11
httptools
pydantic
python-dotenv
PyYAML
starlette<0.14
typing-extensions
uvicorn
uvloop
watchgod
websockets
python-multipart
aiofiles
pyjwt
requests
gunicorn
